feat: intermediate structure
