+++ .pre-commit-config.yaml
@@ -7,14 +7,14 @@ repos:
     hooks:
       - id: ruff
         args: [ --fix, --unsafe-fixes ]
-        exclude: ^(alembic/|.*\.md$|grpc_server/|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
+        exclude: ^(alembic/|.*\.md$|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
 
   - repo: https://github.com/psf/black
     rev: 24.3.0
     hooks:
       - id: black
         language_version: python3
-        exclude: ^(alembic/|.*\.md$|grpc_server/|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
+        exclude: ^(alembic/|.*\.md$|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
 
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: "v5.0.0"
@@ -27,4 +27,4 @@ repos:
       - id: trailing-whitespace
         exclude: ^(alembic/|.*\.md$|grpc_server/|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
       - id: end-of-file-fixer
-        exclude: ^(alembic/|.*\.md$|grpc_server/|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
+        exclude: ^(alembic/|.*\.md$|.*\.pb2\.py$|.*\.pb2_grpc\.py$)
