+++ src/aibolit/grpc_service/grpc_client.py
@@ -2,12 +2,12 @@ from datetime import datetime
 from google.protobuf.timestamp_pb2 import Timestamp
 import grpc
 
-from grpc_server import medications_pb2, medications_pb2_grpc, users_pb2, users_pb2_grpc
+from aibolit.grpc_service.generated import schedule_pb2, schedule_pb2_grpc, user_pb2, user_pb2_grpc
 
 
 async def run():
     async with grpc.aio.insecure_channel('localhost:50051') as channel:
-        stub = medications_pb2_grpc.MedicationServiceStub(channel)
+        stub = schedule_pb2_grpc.SchedulesServiceStub(channel)
         print("1. Create Schedule")
         print("2. Get All Schedules")
         print("3. Get User Schedule")
@@ -19,36 +19,36 @@ async def run():
         if rpc_call == "1":
             start_date = Timestamp()
             start_date.FromDatetime(datetime.now())
-            request = medications_pb2.CreateScheduleRequest(  # type: ignore
+            request = schedule_pb2.CreateScheduleRequest(  # type: ignore
                 user_id=1, medication_name="Test Medication", start_date=start_date, frequency=15, duration_days=10
             )
             reply = await stub.CreateSchedule(request)
             print("Create Schedule Response Received:")
             print(reply)
         if rpc_call == "2":
-            request = medications_pb2.GetAllSchedulesRequest(user_id=1)  # type: ignore
+            request = schedule_pb2.GetAllSchedulesRequest(user_id=1)  # type: ignore
             reply = await stub.GetAllSchedules(request)
             print("Get All Schedules Response Received:")
             print(reply)
         if rpc_call == "3":
-            request = medications_pb2.GetUserScheduleRequest(user_id=1, schedule_id=2)  # type:ignore
+            request = schedule_pb2.GetUserScheduleRequest(user_id=1, schedule_id=2)  # type:ignore
             reply = await stub.GetUserSchedule(request)
             print("Get User Schedule Response Received.")
             print(reply)
 
         if rpc_call == "4":
-            request = medications_pb2.GetUserNextTakingsRequest(user_id=1)  # type: ignore
+            request = schedule_pb2.GetUserNextTakingsRequest(user_id=1)  # type: ignore
             reply = await stub.GetUserNextTakings(request)
             print("Get User Next Takings Response Received.")
             print(reply)
-        stub = users_pb2_grpc.UserServiceStub(channel)
+        stub = user_pb2_grpc.UserServiceStub(channel)
         if rpc_call == "5":
-            request = users_pb2.CreateUserRequest()
+            request = user_pb2.CreateUserRequest()
             reply = await stub.CreateUser(request)
             print("Create User Response Received.")
             print(reply)
         if rpc_call == "6":
-            request = users_pb2.GetAllUsersRequest()
+            request = user_pb2.GetAllUsersRequest()
             reply = await stub.GetUsers(request)
             print("Get All Users Response Received.")
             print(reply)
